Restructure utils.pyro
