+++ scripts/pyro_name_server.py
@@ -7,7 +7,7 @@ POCS.
 """
 import argparse
 
-from huntsman.utils.pyro import run_name_server
+from huntsman.utils.pyro.name_server import run_name_server
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
