+++ scripts/pyro_test_server.py
@@ -2,10 +2,10 @@
 """
 Script to run a Pyro test server.
 """
-from huntsman.utils.pyro import run_test_server
+from huntsman.utils.pyro.test_server import run_test_server
 
 #==============================================================================
 
 if __name__ == "__main__":
-    #Run the test server
+    # Run the test server
     run_test_server()
